requirements.txt
fastapi
uvicorn
openai
python-dotenv
streamlit
requests
python-multipart
